fix authenticateアプリケーションのユニットテストを各モジュール毎の個別ファイルに分割
